Conflicts:
	nbproject/project.properties
	src/melt/DAO/MCQ_DAO.java
	src/melt/DAO/QuestionDAO.java
	src/melt/DAO/Question_DAO.java
	src/melt/Main.java
	src/melt/Model/MCQ.java
	src/melt/Model/Question.java
	src/melt/Util/SQLHelper.java
	src/melt/View/AddExam.java
	src/melt/View/AddSection.java
	src/melt/View/ChooseQuestionsPanel.java
	src/melt/View/Exam.java
	src/melt/View/ExamOverview.java
	src/melt/View/MCQPanel.java
	src/melt/View/SectionPanel.java
	src/melt/View/SettingExam.java
	src/melt/View/SubsectionPanel.java

+++ src/melt/View/Exam.java
@@ -9,7 +9,6 @@ import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.FlowLayout;
-import java.awt.GridLayout;
 import java.awt.Toolkit;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -26,6 +25,7 @@ import javax.swing.*;
 import melt.DAO.MCQ_DAO;
 import melt.DAO.Question_DAO;
 import melt.DAO.Section_DAO;
+import melt.DAO.SubsectionQuestion_DAO;
 import melt.DAO.Subsection_DAO;
 import melt.Model.*;
 
@@ -51,7 +51,7 @@ public class Exam extends JFrame implements ActionListener,WindowListener{
     public Exam(int exam_ID) {
         this.exam_ID=exam_ID;
         //this.setLocationRelativeTo(null);  //make window in the center of desktop
-        setTitle("MELTSystem--Test");
+        setTitle("MELTSystem");
         setDefaultCloseOperation(DO_NOTHING_ON_CLOSE);
         Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
         double width = screenSize.getWidth();
@@ -60,7 +60,7 @@ public class Exam extends JFrame implements ActionListener,WindowListener{
         setContentPane(getGUI());
         //MenuBar
         JMenuBar jMenuBar1=new JMenuBar();
-        JMenu jMenu1=new JMenu("ddd");
+        JMenu jMenu1=new JMenu("Set the exam, by adding Sections, Subsections and Questions");
         jMenuBar1.add(jMenu1);
         getRootPane().setMenuBar(jMenuBar1);
         
@@ -124,7 +124,7 @@ public class Exam extends JFrame implements ActionListener,WindowListener{
         p5.add(p4);
         contentPanel=new JPanel();
         //contentPanel.setBackground(Color.red);
-        contentPanel.setLayout(new BorderLayout());
+        contentPanel.setLayout(new FlowLayout());
         p5.add(contentPanel);
         
         //set the background color of the leftpart
@@ -133,7 +133,7 @@ public class Exam extends JFrame implements ActionListener,WindowListener{
         p4.setBackground(new Color(153, 153, 153));
         
         p4.setBorder(javax.swing.BorderFactory.createTitledBorder(javax.swing.BorderFactory.createTitledBorder(""), "Exam 1"));
-
+        
         return p5;
     }
 
@@ -152,7 +152,7 @@ public class Exam extends JFrame implements ActionListener,WindowListener{
             Section section=sections.get(i-1);
             String sectionTimeLimit=new SimpleDateFormat("HH:mm:ss").format(section.getTimeLimit());
             
-            sectionLabels[i - 1] = new JLabel("Section" + i + "  " + section.getSection_Name());
+            sectionLabels[i - 1] = new JLabel("Section " + i + "  " + section.getSection_Name());
             //Edit Button
             
             //Checkboxes
@@ -255,39 +255,56 @@ public class Exam extends JFrame implements ActionListener,WindowListener{
             contentPanel.setLayout(new FlowLayout());
             contentPanel.add(temp);
             contentPanel.revalidate();
+            for (int i = 0; i < checkbox.length; i++) {
+                checkbox[i].setSelected(false);
+                            checkbox[i].setEnabled(true);
+                        }
             //contentPanel.repaint();
             
         }else if (((JButton)e.getSource()).getText().equals("Delete")) {
             //set section_ID
-            int section_ID=Integer.parseInt(((JButton)e.getSource()).getName());
+            if(checkbox.length > 0)
+            {
+            int section_ID = 0;
+            for (int i = 0; i < checkbox.length; i++) {
+                if (checkbox[i].isSelected()) {
+                    section_ID = Integer.parseInt((checkbox[i].getName()));
+                    //get settingExamPanel and add to right panel
             //create and intialize the DAOs
-            MCQ_DAO mcq_DAO=new MCQ_DAO();
-            Question_DAO question_DAO=new Question_DAO();
+            //MCQ_DAO mcq_DAO=new MCQ_DAO();
+            //Question_DAO question_DAO=new Question_DAO();
+            SubsectionQuestion_DAO subSectionQuestion_DAO=new SubsectionQuestion_DAO();
             Subsection_DAO subsection_DAO=new Subsection_DAO();
             Section_DAO section_DAO=new Section_DAO();
         
             //update the info of question and MCQ then delete the relative data in section and subsection 
-            mcq_DAO.cancelRWithSubSec(section_ID);
-            question_DAO.cancelRWithSubSec(section_ID);
+            //mcq_DAO.cancelRWithSubSec(section_ID);
+            //question_DAO.cancelRWithSubSec(section_ID);
+            subSectionQuestion_DAO.cancelRWithSubSec(section_ID);
             subsection_DAO.delete("Section_ID='"+section_ID+"'");
             section_DAO.delete("Section_ID='"+section_ID+"'");
             
             p2.removeAll();
             
             setP2();
+                }
+            }
+            }
+        
             
             
             
         }else if(((JButton)e.getSource()).getText().equals("Edit")){  //Edit Button
            // this.dispose();
+            if(checkbox.length > 0)
+            {
             int section_ID = 0;
             for (int i = 0; i < checkbox.length; i++) {
                 if (checkbox[i].isSelected()) {
                     section_ID = Integer.parseInt((checkbox[i].getName()));
-                }
-            }
-          
-            //get settingExamPanel and add to right panel
+
+                    //get settingExamPanel and add to right panel
+
             SettingExam settingExam=new SettingExam(section_ID);
             contentPanel.removeAll();
             contentPanel.setLayout(new BorderLayout());
@@ -295,6 +312,12 @@ public class Exam extends JFrame implements ActionListener,WindowListener{
             contentPanel.revalidate();
             //contentPanel.repaint();
             //settingExam.setVisible(true);
+                }
+            }
+          
+            
+            
+        }
             
         }
     }
